skills/mycroft-wiki.mycroftai/requirements.txt
pymediawiki~=0.7.0
requests